Conflicts:
	JMX.xcodeproj/project.pbxproj
	entities/JMXAudioSpectrumAnalyzer.mm

+++ HID Utilities/HID_Utilities.c
@@ -281,7 +281,7 @@ void HIDRebuildDevices( void ) {
 
 UInt32 HIDCountDevices( void )
 {
-	return CFArrayGetCount( gDeviceCFArrayRef );
+	return (UInt32)CFArrayGetCount( gDeviceCFArrayRef );
 }
 
 // ---------------------------------
